@@ -4,20 +4,31 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   ptv_api_client:
-    git: https://github.com/nickmeinhold/ptv_api_client.git
+    git:
+      url: https://github.com/nickmeinhold/ptv_api_client.git
+      ref: built_value
+      
   geolocator: any
   flutter_redux: any
+  redux_remote_devtools: any
+  built_collection: '>=2.0.0 <5.0.0'
+  built_value: '>=5.5.5 <7.0.0'
+  flutter_staggered_animations: "^0.1.1"
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.0.0
+  build_runner: ^1.0.0
+  built_value_generator: ^6.7.0
+  pedantic: ^1.4.0
+  quiver: '>=0.21.0 <3.0.0'
+  test: ^1.0.0
 
 flutter:
   uses-material-design: true
